refactor(handoff): extract remote transport to shared library
Lift the push/pull/list/bootstrap surface out of the 1389-line
bin/dotclaude-handoff.mjs into a new module at
plugins/dotclaude/src/lib/handoff-remote.mjs. This is a pure refactor:
help output and all CLI behavior are byte-identical, and the existing
240 bats tests + 284 vitest tests continue to pass unchanged.

The library now owns:
- session extraction (extractMeta / extractPrompts / extractTurns)
- digest rendering (renderHandoffBlock, mechanicalSummary, nextStepFor)
- URL / path primitives (validateTransportUrl, slugify, v2BranchName,
  monthBucket, projectSlugFromCwd, isRepoMissingError)
- metadata encode / decode
- persisted-env loading and interactive bootstrap (loadPersistedEnv,
  bootstrapTransportRepo, requireTransportRepo{,Strict})
- remote I/O (pushRemote, pullRemote, fetchRemoteBranch,
  listRemoteCandidates, enrichWithDescriptions, matchesQuery)

The bin keeps argv parsing, the resolver/collision-prompt bridge,
local enumeration (CLI_LAYOUTS + listLocalSessions), search,
host detection, and the main dispatch switch. It re-exports the
names the existing vitest suites already import from the bin so
those tests don't need touching.

Scope deviation worth flagging: the plan originally said renderers
stay in the bin. I moved them into the library because pushRemote
depends on them — keeping pushRemote's signature intact was cheaper
than threading the digest in from the caller. The bin re-imports
renderers from the library for the digest/file/bare-query verbs.

Issue #91 Gap 1 — step 2 of docs/plans/handoff-issue-rollout.md.
Sets up the single seam subsequent gaps (#91 Gaps 2-8, #90 Gaps
2-4, #85) will modify instead of re-editing the bin each time.

Spec ID: dotclaude-core
